+++ joint_genotyping_record.cpp
@@ -672,7 +672,7 @@ void JointGenotypingRecord::process_read(AugmentedBAMRecord& as, int32_t sampleI
 		allele = 1;
 	      }
 	      else {
-		q = abs(len-aug_ref[i].size())*30;
+		q = abs((int32_t)len-(int32_t)aug_ref[i].size())*30;
 		allele = -1;
 	      }
 	      
@@ -695,7 +695,7 @@ void JointGenotypingRecord::process_read(AugmentedBAMRecord& as, int32_t sampleI
 		allele = 1;
 	      }
 	      else {
-		q = abs(len-aug_ref[i].size())*30;
+		q = abs((int32_t)len-(int32_t)aug_ref[i].size())*30;
 		allele = -1;
 	      }
 	      cycle = strand == 'F' ? (rpos0+1) : (rlen - rpos0);
